hifive1: Move MCU devices from bsp to mcu
Timers, uart and spi belong to MCU not bsp so they are
not moved to correct place.
Those can be created in hal_bsp.c by calling fe310_periph_create().

HIFIVE1 did not have SPI2 available on board but fe310 could have
it.
